Fixed crash on startup, menu glitch, memory leak
The app does not crash on startup any more. This was caused by an early window push.

A glitch where the menu would not be displayed correctly every even time it was opened is now fixed. The memory leak associated with this has also been patched.

The app no deallocates all memory before terminating.